Changes to fix EddyQC in minified container
While the purpose of this change is to create a new minified version of FSL, other changes were required to actually facilitate such.
For FSL itself, in addition to an update of version number, the Python installation script is now used, and the executables generated in $FSLDIR/share/fsl/bin are used.
In addition, the generated tarball needs to be inclusive of softlinks, given that the eddy_quad executable shebang points to a softlink.